Fixed visualization
Had to rollback a few scripts that didnt wofk with the new vector/scalar formatting. I think i need to account for a fields chance to go from -1 to 1 or 0 to 1.

Advection and Diffusion seem to be working well. Visuals are in reach, as the fluid acts as a contained ball, fading away depending on viscosity/material/velocity dissipation. more sliders probably needed.