Rolled back, but updated
rolled back but integrated a lot of the new vector system into the old system. the color adding is just staying constant as there is no need to convert it to a certain style just to immediately unconvert again.

Working on buoyancy, trying to get that turbulent trail. That seems to be the last major effect missing before most of the effects can be checked off.