Found a scaling issue
The add density and add velocity functions are being drawn very small in the top left, and normal sized in the bottom right. this is affecting how the entire field changes, and the boundary conditions are also too messy. going back and trying to fix this core scale issue before fixing the scale in the shaders.