New velocity ield
Reworked the way velocity is added to the fluid, being a dynamic thing depending on mouse or other game object input. this new system works with advection and diffusion, and is starting to look nice.

The divergence and curl steps are broken, I think it has something to do with the way signed scalar fields are stored.